# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type PaystackOrderItemFlavor = typeof PaystackOrderItemFlavor[keyof typeof PaystackOrderItemFlavor];


export const PaystackOrderItemFlavor = {
  'salted-caramel': 'salted-caramel',
  'peanut-brittle': 'peanut-brittle',
  'caramel-cheese': 'caramel-cheese',
} as const;

export interface PaystackOrderItem {
  flavor: PaystackOrderItemFlavor;
  /**
     * @minimum 1
     * @maximum 20
     */
  quantity: number;
}

export type PaystackCheckoutInputState = typeof PaystackCheckoutInputState[keyof typeof PaystackCheckoutInputState];


export const PaystackCheckoutInputState = {
  abuja: 'abuja',
} as const;

export interface PaystackCheckoutInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  email: string;
  /**
     * @minLength 7
     * @maxLength 32
     */
  phone: string;
  /**
     * @minLength 5
     * @maxLength 500
     */
  address: string;
  state: PaystackCheckoutInputState;
  /**
     * @minItems 1
     * @maxItems 3
     */
  items: PaystackOrderItem[];
  callbackUrl: string;
}

export interface PaystackCheckoutSession {
  authorizationUrl: string;
  reference: string;
}

export interface PaystackPaymentConfirmation {
  verified: boolean;
  customerName: string;
  totalNaira: number;
}

export interface AdminOrder {
  reference: string;
  customerName: string;
  /** @nullable */
  email: string | null;
  phone: string;
  address: string;
  state: string;
  items: string;
  totalNaira: number;
  /** @nullable */
  paidAt: string | null;
  /** @nullable */
  deliveredAt: string | null;
}

export interface AdminDeliveryUpdate {
  delivered: boolean;
}

export interface AdminDeliveryStatus {
  reference: string;
  /** @nullable */
  deliveredAt: string | null;
}

export interface AdminOrderPage {
  orders: AdminOrder[];
  page: number;
  hasMore: boolean;
}

export type ListAdminOrdersParams = {
/**
 * @minimum 1
 * @maximum 10000
 */
page?: number;
};