# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: payments
    description: Checkout and payment verification
  - name: admin
    description: Private shop management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /paystack/checkout:
    post:
      operationId: createPaystackCheckout
      tags: [payments]
      summary: Start a Paystack checkout for a popcorn order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaystackCheckoutInput"
      responses:
        "200":
          description: Hosted Paystack checkout created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaystackCheckoutSession"
        "400":
          description: Invalid order or unavailable delivery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Paystack is temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Paystack is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /paystack/verify/{reference}:
    get:
      operationId: verifyPaystackTransaction
      tags: [payments]
      summary: Confirm a Paystack transaction before showing order success
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
            minLength: 3
            maxLength: 100
      responses:
        "200":
          description: Payment was successful and matches the order total
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaystackPaymentConfirmation"
        "400":
          description: Invalid transaction reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Payment is not confirmed or does not match the order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Paystack verification is temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Paystack is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List confirmed Burst Popcorn Paystack orders for the shop admin
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 1
      responses:
        "200":
          description: Confirmed orders from a page of Paystack transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderPage"
        "401":
          description: Sign in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Not a shop admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Could not read Paystack transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Paystack or admin access is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/orders/{reference}/delivery:
    patch:
      operationId: updateAdminOrderDelivery
      tags: [admin]
      summary: Mark a confirmed Burst Popcorn order delivered or undo delivery
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDeliveryUpdate"
      responses:
        "200":
          description: Updated delivery status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDeliveryStatus"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Sign in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Not a shop admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Confirmed Burst order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Could not verify payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Admin access or delivery storage is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
    PaystackOrderItem:
      type: object
      required: [flavor, quantity]
      properties:
        flavor:
          type: string
          enum: [salted-caramel, peanut-brittle, caramel-cheese]
        quantity:
          type: integer
          minimum: 1
          maximum: 20
    PaystackCheckoutInput:
      type: object
      required: [name, email, phone, address, state, items, callbackUrl]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
        phone:
          type: string
          minLength: 7
          maxLength: 32
        address:
          type: string
          minLength: 5
          maxLength: 500
        state:
          type: string
          enum: [abuja]
        items:
          type: array
          minItems: 1
          maxItems: 3
          items:
            $ref: "#/components/schemas/PaystackOrderItem"
        callbackUrl:
          type: string
          format: uri
    PaystackCheckoutSession:
      type: object
      required: [authorizationUrl, reference]
      properties:
        authorizationUrl:
          type: string
          format: uri
        reference:
          type: string
    PaystackPaymentConfirmation:
      type: object
      required: [verified, customerName, totalNaira]
      properties:
        verified:
          type: boolean
        customerName:
          type: string
        totalNaira:
          type: integer
    AdminOrder:
      type: object
      required: [reference, customerName, email, phone, address, state, items, totalNaira, paidAt, deliveredAt]
      properties:
        reference:
          type: string
        customerName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: string
        address:
          type: string
        state:
          type: string
        items:
          type: string
        totalNaira:
          type: integer
        paidAt:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
          format: date-time
    AdminDeliveryUpdate:
      type: object
      required: [delivered]
      properties:
        delivered:
          type: boolean
    AdminDeliveryStatus:
      type: object
      required: [reference, deliveredAt]
      properties:
        reference:
          type: string
        deliveredAt:
          type: ["string", "null"]
          format: date-time
    AdminOrderPage:
      type: object
      required: [orders, page, hasMore]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/AdminOrder"
        page:
          type: integer
        hasMore:
          type: boolean