# lib/api-zod/src/generated/types/adminDeliveryStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminDeliveryStatus {
  reference: string;
  /** @nullable */
  deliveredAt: Date | null;
}