# lib/api-zod/src/generated/types/adminOrder.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminOrder {
  reference: string;
  customerName: string;
  /** @nullable */
  email: string | null;
  phone: string;
  address: string;
  state: string;
  items: string;
  totalNaira: number;
  /** @nullable */
  paidAt: string | null;
  /** @nullable */
  deliveredAt: Date | null;
}