# lib/api-zod/src/generated/types/adminOrderPage.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminOrder } from './adminOrder';

export interface AdminOrderPage {
  orders: AdminOrder[];
  page: number;
  hasMore: boolean;
}